Create BinarySearch
Given a sorted array of integers, check if the given number is present or not.

Example 1:
Input array = {-2, 4, 7, 9, 11}
Number to be searched = 9

Output = true


Example 2:
Input array = {-2, 4, 7, 9, 11}
Number to be searched = 6

Output = false

Code:
public boolean search(int[] arr, int num)
{
     // Code goes here
}

Tips:
It is easy to traverse the entire array and check if the number is present or not. But the time complexity of the solution will be O(n). Can we do better? Like a solution with complexity of O(logn)? (Hint: Binary search)